+++ packages/themes/CHANGELOG.md
@@ -9,6 +9,7 @@
 ### Changed
 
 - Updated component shadows using theme.
+- Added `textFieldContainer` token for filled text field background styling.
 
 # v9.0.1 (October 16, 2025)